Create vpEigenConversion.cpp file to contain definition of non template eigen2visp functions. Otherwise, building visp_java module lead to duplicate these functions in listconverters.cpp.o and visp_java.cpp.o
